Moved the numerically stable general exponential into the main library and out of tools and swapped it with rpow
